api/requirements.txt
Django==1.8.2
django-filter==0.10.0
djangorestframework==3.1.2
Markdown==2.6.2
requests==2.7.0
wheel==0.24.0
Django==1.8.2
django-cors-headers==1.1.0
django-filter==0.10.0
djangorestframework==3.1.2
Markdown==2.6.2
requests==2.7.0
wheel==0.24.0
